+++ packages/core/field-number/examples/00-basic-example.js
@@ -1,7 +1,7 @@
 // @flow
 import React, { PureComponent } from 'react';
 import { Form, formDefaultProps } from '@uidu/form';
-import { inputDefaultProps } from '@uidu/field-base';
+import { inputDefaultProps } from '@uidu/field-base/examples-utils';
 
 import TextField from '../src';