+++ packages/core/select/examples/14-with-group-heading.js
@@ -1,8 +1,7 @@
 // @flow
 import React from 'react';
 import { Form, formDefaultProps } from '@uidu/form';
-import { inputDefaultProps } from '@uidu/field-base';
-
+import { inputDefaultProps } from '@uidu/field-base/examples-utils';
 import Select from '../src';
 
 const GROUP_OPTIONS = [